requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlite-utils==3.35.2
pydantic==2.5.0
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiofiles==23.2.1

# Data Analysis & Visualization
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
scipy==1.11.4
scikit-learn==1.3.2

# Market Data APIs
yfinance==0.2.28
alpha_vantage==2.3.1
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
python-dotenv==1.0.0

# Database
alembic==1.13.0